+++ src/painting/ClassDecoration.ts
@@ -4,29 +4,7 @@ class ClassDecoration{
     color?: string;
     display?: string;
     flex?: string;
-    // background?: string;
-    // background?: string;
-    // background?: string;
-    // background?: string;
-    // background?: string;
-    // background?: string;
-    // background?: string;
-    // background?: string;
-    // background?: string;
-    // background?: string;
-    // background?: string;
-    // background?: string;
-    // background?: string;
-    // background?: string;
-    // background?: string;
-    // background?: string;
-    // background?: string;
-    // background?: string;
-    // background?: string;
-    // background?: string;
-    // background?: string;
-    // background?: string;
-    // background?: string;
+
 }
 
 export{