+++ src/rendering/flex.ts
@@ -14,7 +14,7 @@ enum MainAxisAlignment {
     * If this value is used in a vertical direction, a [VerticalDirection] must be
     * available to determine if the start is the top or the bottom.
     * */
-    start = "justify-start",
+    start = "flex-start",
   
     /** Place the children as close to the end of the main axis as possible.
     *
@@ -24,23 +24,23 @@ enum MainAxisAlignment {
     * If this value is used in a vertical direction, a [VerticalDirection] must be
     * available to determine if the end is the top or the bottom.
     * */
-    end = "justify-end",
+    end = "flex-end",
   
     /** Place the children as close to the middle of the main axis as possible.*/
-    center = "justify-center",
+    center = "center",
   
     /** Place the free space evenly between the children.*/
-    spaceBetween = "justify-between",
+    spaceBetween = "space-between",
   
     /** Place the free space evenly between the children as well as half of that
     * space before and after the first and last child.
     * */
-    spaceAround = "justify-around",
+    spaceAround = "space-around",
   
     /** Place the free space evenly between the children as well as before and
     * after the first and last child.
     * */
-    spaceEvenly = "justify-evenly",
+    spaceEvenly = "space-evenly",
   }
   enum CrossAxisAlignment {
     /** Place the children as close to the start of the main axis as possible.
@@ -51,7 +51,7 @@ enum MainAxisAlignment {
     * If this value is used in a vertical direction, a [VerticalDirection] must be
     * available to determine if the start is the top or the bottom.
     * */
-    start = "items-start",
+    start = "start",
   
     /** Place the children as close to the end of the main axis as possible.
     *
@@ -61,15 +61,15 @@ enum MainAxisAlignment {
     * If this value is used in a vertical direction, a [VerticalDirection] must be
     * available to determine if the end is the top or the bottom.
     * */
-    end = "items-end",
+    end = "end",
   
     /** Place the children as close to the middle of the main axis as possible.*/
-    center = "items-center",
+    center = "center",
   
     /** Place the free space evenly between the children as well as half of that
     * space before and after the first and last child.
     * */
-    spaceAround = "base-line",
+    spaceAround = "stretch",
   
     /** Place the free space evenly between the children as well as before and
     * after the first and last child.