+++ src/widgets/framework.ts
@@ -70,10 +70,10 @@ export default class Widget {
     boxDecoration?: _BoxDecoration;
     css?: CssProperties;
 
-    margin?:EdgeInsetsGeometry,
-    padding?:EdgeInsetsGeometry,
+    margin?: EdgeInsetsGeometry;
+    padding?: EdgeInsetsGeometry;
   }) {
-    this.css = new CssProperties({});
+    this.css = new CssProperties({ ...args.css?._props });
 
     this.key = args.key;
     this.child = args.child;
@@ -98,23 +98,20 @@ export default class Widget {
 
     this.classes = args.classes;
 
-      let paddings = args.padding?.arrayed_value("padding");
-        if(paddings){
-
-          this.css!._props.paddingLeft = paddings![0];
-          this.css!._props.paddingTop = paddings![1];
-          this.css!._props.paddingRight = paddings![2];
-          this.css!._props.paddingBottom = paddings![3];
-           
-        }
-        let margins = args.margin?.arrayed_value("margin");
-        if(margins){
-          this.css!._props.marginLeft = margins![0];
-          this.css!._props.marginTop = margins![1];
-          this.css!._props.marginRight = margins![2];
-          this.css!._props.marginBottom = margins![3];
-        }
-        
+    let paddings = args.padding?.arrayed_value("padding");
+    if (paddings) {
+      this.css!._props.paddingLeft = paddings![0];
+      this.css!._props.paddingTop = paddings![1];
+      this.css!._props.paddingRight = paddings![2];
+      this.css!._props.paddingBottom = paddings![3];
+    }
+    let margins = args.margin?.arrayed_value("margin");
+    if (margins) {
+      this.css!._props.marginLeft = margins![0];
+      this.css!._props.marginTop = margins![1];
+      this.css!._props.marginRight = margins![2];
+      this.css!._props.marginBottom = margins![3];
+    }
 
     if (args.boxDecoration?.color)
       this.css!._props.backgroundColor = args.boxDecoration?.color ?? "";