+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+# v0.2.2-alpha-1 (2022-06-13)
+This version comes with Some Features, Updates and Enhancements
 ## Feature
 * Added `Radius` Class.
 * Added `BorderRadiusGeometry` Class.